backend-services/services/ai-optimizer/src/main.py
from flask import Flask, jsonify
from flask_cors import CORS
import random
import os
import json
from google.cloud import pubsub_v1
from google.cloud import storage
from google.cloud import bigquery
from google.cloud import bigtable
from google.cloud import secretmanager
import psycopg2
import redis

app = Flask(__name__)
CORS(app)

# GCP Clients (with fallback for local development)
project_id = os.getenv('PROJECT_ID', 'alpha-orion')
try:
    publisher = pubsub_v1.PublisherClient()
    storage_client = storage.Client()
    bigquery_client = bigquery.Client()
    bigtable_client = bigtable.Client(project=project_id)
    secret_client = secretmanager.SecretManagerServiceClient()
    gcp_available = True
    print("GCP services initialized successfully")
except Exception as e:
    print(f"GCP services not available (expected in local dev): {e}")
    publisher = None
    storage_client = None
    bigquery_client = None
    bigtable_client = None
    secret_client = None
    gcp_available = False

# Connections
db_conn = None
redis_conn = None

def get_db_connection():
    global db_conn
    if db_conn is None:
        db_url = os.getenv('DATABASE_URL')
        db_conn = psycopg2.connect(db_url)
    return db_conn

def get_redis_connection():
    global redis_conn
    if redis_conn is None:
        redis_url = os.getenv('REDIS_URL')
        redis_conn = redis.from_url(redis_url)
    return redis_conn

@app.route('/optimize', methods=['GET'])
def optimize():
    # Mock AI optimization
    optimizations = {
        'strategy': 'Arbitrage',
        'parameters': {
            'threshold': random.uniform(0.01, 0.05),
            'maxSlippage': random.uniform(0.001, 0.01)
        },
        'expectedReturn': random.uniform(1.0, 5.0)
    }
    return jsonify(optimizations)

@app.route('/health', methods=['GET'])
def health():
    return jsonify({'status': 'ok'})

if __name__ == '__main__':
    app.run(host='0.0.0.0', port=8080)
